VictimProjects/SmartLight/web_dashboard/backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
paho-mqtt==1.6.1
gunicorn==21.2.0
python-dotenv==0.19.0
